+++ tests/index.test.js
@@ -1,2 +1,3 @@
 /* eslint-disable no-unused-vars */
 import './assistant.test';
+import './iframe.test';